feat(app): add a modal component
